requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.1
matplotlib==3.8.2
python-dotenv==1.0.0
numpy==1.26.3
python-bidi==0.4.2 